+++ hutool-core/src/main/java/cn/hutool/core/net/url/UrlBuilder.java
@@ -544,12 +544,7 @@ public final class UrlBuilder implements Builder<String> {
 	 */
 	public URI toURI() {
 		try {
-			return new URI(
-					getSchemeWithDefault(),
-					getAuthority(),
-					getPathStr(),
-					getQueryStr(),
-					getFragmentEncoded());
+			return toURL().toURI();
 		} catch (URISyntaxException e) {
 			return null;
 		}